Split Frame -> WinFrame.
Some of the functions in WinFrame.cpp will probably become part of the "interface".
E.g.: FrameRefreshStatus, FrameUpdateApple2Type, FrameDrawDiskLEDS.
