+++ source/Configuration/PageConfig.cpp
@@ -27,7 +27,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "PropertySheetHelper.h"
 
 #include "../AppleWin.h"
-#include "../Frame.h"
+#include "../Windows/WinFrame.h"
 #include "../Registry.h"
 #include "../SerialComms.h"
 #include "../Windows/WinVideo.h"
