+++ source/Configuration/PageDisk.cpp
@@ -29,7 +29,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "../AppleWin.h"
 #include "../CardManager.h"
 #include "../Disk.h"	// Drive_e, Disk_Status_e
-#include "../Frame.h"
+#include "../Windows/WinFrame.h"
 #include "../Registry.h"
 #include "../resource/resource.h"
