+++ source/Debugger/Debugger_Assembler.cpp
@@ -31,7 +31,6 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "Debug.h"
 
 #include "../CPU.h"
-#include "../Frame.h"
 #include "../Memory.h"
 
 #define DEBUG_ASSEMBLER 0
