+++ source/Debugger/Debugger_Commands.cpp
@@ -27,7 +27,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "Debug.h"
 
-#include "../Frame.h"
+#include "../Windows/WinFrame.h"
 
 // Commands _______________________________________________________________________________________
