+++ source/Debugger/Debugger_Display.cpp
@@ -35,6 +35,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "../AppleWin.h"
 #include "../CPU.h"
 #include "../Frame.h"
+#include "../Windows/WinFrame.h"
 #include "../LanguageCard.h"
 #include "../Memory.h"
 #include "../Mockingboard.h"
