+++ source/Disk.cpp
@@ -37,13 +37,13 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "AppleWin.h"
 #include "CPU.h"
 #include "DiskImage.h"
-#include "Frame.h"
 #include "Log.h"
 #include "Memory.h"
 #include "Registry.h"
 #include "SaveState.h"
 #include "Video.h"
 #include "Windows/WinVideo.h"
+#include "Windows/WinFrame.h"
 #include "YamlHelper.h"
 
 #include "../resource/resource.h"
