+++ source/Keyboard.cpp
@@ -30,7 +30,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "Keyboard.h"
 #include "AppleWin.h"
-#include "Frame.h"
+#include "Windows/WinFrame.h"
 #include "Pravets.h"
 #include "Tape.h"
 #include "YamlHelper.h"
