+++ source/ParallelPrinter.cpp
@@ -30,7 +30,6 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "ParallelPrinter.h"
 #include "AppleWin.h"
-#include "Frame.h"	// g_hFrameWindow
 #include "Memory.h"
 #include "Registry.h"
 #include "YamlHelper.h"
