+++ source/SerialComms.cpp
@@ -38,7 +38,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "SerialComms.h"
 #include "AppleWin.h"
 #include "CPU.h"
-#include "Frame.h"
+#include "Windows/WinFrame.h"
 #include "Log.h"
 #include "Memory.h"
 #include "YamlHelper.h"
