+++ source/Windows/WinVideo.cpp
@@ -29,6 +29,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "Windows/WinVideo.h"
+#include "Windows/WinFrame.h"
 #include "AppleWin.h"
 #include "Video.h"
 #include "CPU.h"
